ref: update results controller methods and add module
